pyproject.toml
[tool.poetry]
name = "dataxml"
version = "0.1.0"
description = ""
authors = ["Chaipat Suwannapoom <bchaipats@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyspark = "^3.4.1"
pytest = "^7.4.0"
pre-commit = "^3.3.3"
pandas = "^2.0.3"
scikit-learn = "^1.3.0"
delta-spark = "^2.4.0"
mlflow = "^2.5.0"
mypy = "^1.4.1"

[tool.poetry.group.github-actions]
# only used in github actions
optional = true
[tool.poetry.group.github-actions.dependencies]
pytest-github-actions-annotate-failures = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
